requirements.txt
numpy>=1.21.0
huggingface-hub>=0.19.0
python-dotenv>=0.19.0
requests>=2.26.0
transformers>=4.30.0
torch>=2.0.0
langchain>=0.0.200
PyGithub>=2.1.1
